+++ packages/apps/ai/sagemaker-mlops-app/lib/sagemaker-mlops.ts
@@ -60,7 +60,7 @@ export class SageMakerMLOpsApp extends MdaaCdkApp {
       codeStarConnection: config.training.codeStarConnection,
       baseJobPrefix: config.training.baseJobPrefix,
       trainingDataPath: config.training.trainingDataPath,
-      codeArtifact: config.training.codeArtifact,
+      buildPolicies: config.training.buildPolicies,
     };
 
     const training = new SageMakerModelTrainingL3Construct(stack, 'training', trainingProps);
@@ -92,7 +92,7 @@ export class SageMakerMLOpsApp extends MdaaCdkApp {
       sourceType: config.deploy.sourceType,
       codeStarConnection: config.deploy.codeStarConnection,
       cdkBootstrapQualifier: config.deploy.cdkBootstrapQualifier,
-      codeArtifact: config.deploy.codeArtifact,
+      buildPolicies: config.deploy.buildPolicies,
     };
 
     new SageMakerModelDeployL3Construct(stack, 'deploy', deployProps);